fix: fix potential mv failure

+++ scripts/system_setup.sh
@@ -19,7 +19,10 @@ sudo systemctl enable -f ly@tty1.service
 
 DOTFILES_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")/.." && pwd)"
 echo "Linking ly configuration..."
-sudo mv /etc/ly/config.ini /etc/ly/config.bak
+
+if [ -f "/etc/ly/config.ini" ]; then
+	sudo mv /etc/ly/config.ini /etc/ly/config.bak
+fi
 sudo stow -d "$DOTFILES_DIR" -R -t / system
 
 echo "Step 3 completed successfully!"
